Move JournalReader and JournalWriter to common
