+++ lrucache/src/jvmMain/kotlin/com/mayakapps/lrucache/Utils.kt
@@ -13,15 +13,3 @@ internal fun File.renameToOrThrow(dest: File, deleteDest: Boolean) {
 internal fun File.deleteOrThrow() {
     if (exists() && !delete()) throw IOException()
 }
-
-
-internal fun DataOutputStream.writeLengthString(string: String) {
-    writeByte(string.length)
-    writeString(string)
-}
-
-internal fun DataOutputStream.writeString(string: String) = write(string.encodeToByteArray())
-
-internal fun DataInputStream.readString() = readString(readByte().toInt())
-internal fun DataInputStream.readString(length: Int) = readBytes(length).decodeToString()
-internal fun DataInputStream.readBytes(count: Int) = ByteArray(count).also { read(it) }
